Clean up log test factories
