+++ src/test/kotlin/ui/logs/ViewProjectHistoryUIControllerTest.kt
@@ -1,7 +1,7 @@
 package ui.logs
 
+import helper.LogTestFactory.LOGS_FOR_PROJECT_1
 import helper.ProjectFactory.PROJECT_1
-import helper.ViewProjectHistoryTestFactory.LOGS_FOR_PROJECT_1
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
